ref: split bench mark <- router
